Refactored members page from users route into smaller chunks (#1044)
* fix/refactored members page from users into smaller chunks


* fix/ellipse comment


---------
