+++ nextjs/src/features/users/components/UtilFunctions.ts
@@ -1,0 +1,10 @@
+export const getStatuses = (status: string): string => {
+  switch (status) {
+    case 'pending':
+      return 'status-pending'
+    case 'accepted':
+      return 'status-accepted'
+    default:
+      return 'status-rejected'
+  }
+}
